fix typo: PATCH -> PLUGINS
After listing plugin with `make; make plugins`, a message is displayed
telling the user to check the `PATCH` section in the `config.mk` file.

I think it was intended to point the user to the `PLUGINS` variable in
that file, which is where plugins are set (there is no mention of PATCH
anywhere in the repository).