stest: fix bugs that deviate from original stest
This fixes two deviations from the C implementation of stest in the
original dmenu repository.

The first deviation is related to the existence test and its -e option.
The original stest always performs an existence test, whether the -e
option is supplied or not. This is surprising behavior and may be a bug
in the original implementation. It's related to the fact that the
original implementation performs an initial stat system call in order to
obtain file information to use for further tests. The original stest
requires that this call succeed, which thus requires that input files
exist. This stat call can be seen in the [first line][1] of test
conditions

    if ((!stat(path, &st) && (FLAG('a') || name[0] != '.')

Notably, this behavior deviates from the behavior of GNU coreutils'
test, which will pass when given a non-existing file if the -e option is
not explicitly provided. Though, in fairness, test is a more general
utility oriented towards shell expressions while stest is specifically
oriented towards files. Prior to this commit, the rust implementation of
stest here was consistent with the behavior of GNU coreutils' test, but
it is now consistent with the behavior of the original stest.

The second deviation is related to how the newer than (-n) and older
than (-o) options behave when they are given a non-existing file as
their argument -- for example, when testing that a stream of files are
each newer than a non-existing file. What's the modification time of a
non-existing file? Certainly it's an edge-case. The original stest will
print an error message with perror, ignore the -n or -o flag, and
continue to test files according to the other options specified. This
behavior is implemented with the following [if statement][2]

    if (!(FLAG(ARGC()) = !stat(file, (ARGC() == 'n' ? &new : &old))))
	perror(file);

which, when unpacked, will disable the FLAG('n') or FLAG('o') options
when the stat system call fails (e.g., with [ENOENT][3]). Prior to this
commit, the rust implementation of stest here would panic when a
non-existing file was passed to either the newer than or older than
option; now it is consistent with the behavior of the original stest.

Notably, the behavior of the newer than option is relied upon in the
dmenu_path shell script! It invokes stest as

    stest -dqr -n "$cache" $PATH

in [the block][4]

    if stest -dqr -n "$cache" $PATH; then
	    stest -flx $PATH | sort -u | tee "$cache"
    else
	    cat "$cache"
    fi

The $cache file, typically ~/.cache/dmenu_run, may or may not exist. If
it does not exist, users of dmenu_path expect the stest call to succeed
and the script to enter the first case in the if statement and recreate
the $cache file. Thus, this behavior is related to the issue raised in
[PR #56][5].

Finally, while the original stest behavior of the newer than option is
consistent with analogous behavior in GNU coreutils' test, the behavior
of the older than option is not! When an older than comparison against a
non-existing file is performed in GNU coreutils' test, the result always
fails. Two analogous examples might be expressed as

    stest -o doesnotexist file

    test doesnotexist -ot file

Assuming file exists, the above stest program always succeeds while the
analogous test program always fails. The approach for both options can
be roughly interpreted as the following for each program

    stest: ignore the option if it's file argument does not exist

    test: the modification time of non-existing files is a lower bound

The behavior of GNU coreutils' test is, naturally, also the result of
how it navigates failing stat system calls. You can inspect the source
and note that test's newer than and older than options guard their
comparisons on a successful return of stat. Exactly which stat call is
used as a guard varies by option -- the left-hand exit code (le) in the
case of a [newer than][6] (-nt) comparison, and the right-hand exit code
(re) in the case of an [older than][7] (-ot) comparison. This leads to
the behavior described above.

This analysis was based on the following commits from each repository

    dmenu-rs ef80ff29907ab700b470f80e49e6c1b2192bb9e9
    dmenu 7ab0cb5ef0e19352fc5d64ae0d57a5cf4540acbf
    coreutils 3564e3a572b64c5384519d735dbbb88690438eae

[1]: https://git.suckless.org/dmenu/file/stest.c.html#l28
[2]: https://git.suckless.org/dmenu/file/stest.c.html#l73
[3]: https://www.man7.org/linux/man-pages/man2/stat.2.html
[4]: https://git.suckless.org/dmenu/file/dmenu_path.html#l9
[5]: https://github.com/Shizcow/dmenu-rs/pull/56
[6]: https://git.savannah.gnu.org/cgit/coreutils.git/tree/src/test.c#n316
[7]: https://git.savannah.gnu.org/cgit/coreutils.git/tree/src/test.c#n345